+++ back-end/src/main/java/com/ssafy/health/dto/response/FacilityLocationResponse.java
@@ -4,8 +4,8 @@ import java.math.BigDecimal;
 
 public record FacilityLocationResponse(
         Long facilityId,
-        BigDecimal latitude,
-        BigDecimal longitude,
+        Double latitude,
+        Double longitude,
         Boolean status
 ) {
 }
